Add track variable and update producer options
The code changes include adding a new variable called "track" of type MediaStreamTrack. It is assigned the value of the first video track of the localStream object. Additionally, the producer options for sendTransport.produce are updated to include the "track" variable along with other parameters. The 'trackended' and 'transportclose' event listeners are also added to the producer object.